saving prog

+++ log_core/include/typedefs.hpp
@@ -1,5 +1,5 @@
 #ifndef TYPEDEFS_HPP
 #define TYPEDEFS_HPP
 typedef unsigned int RoomID;
-typedef unsigned int Timestamp;
+typedef unsigned long Timestamp;
 #endif
